+++ app/Actions/FrontArticle/LatestAction.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace App\Actions\FrontArticle;
 
 use App\Models\Article;
-use App\Repositories\ArticleRepository;
+use App\Repositories\Article\FrontArticleRepository;
 use Illuminate\Pagination\LengthAwarePaginator;
 
 class LatestAction
 {
-    public function __construct(private ArticleRepository $articleRepository) {}
+    public function __construct(private FrontArticleRepository $articleRepository) {}
 
     /**
      * @return LengthAwarePaginator<int, Article>
