+++ app/Http/Controllers/Mypage/Article/IndexController.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Mypage\Article;
 
 use App\Http\Resources\Frontend\UserShow;
-use App\Repositories\ArticleRepository;
+use App\Repositories\Article\MypageArticleRepository;
 use App\Services\Front\MetaOgpService;
 use Illuminate\Contracts\View\View;
 use Illuminate\Routing\Controller;
@@ -14,7 +14,7 @@ use Illuminate\Support\Facades\Auth;
 class IndexController extends Controller
 {
     public function __construct(
-        private readonly ArticleRepository $articleRepository,
+        private readonly MypageArticleRepository $articleRepository,
         private readonly MetaOgpService $metaOgpService,
     ) {}
