+++ app/Http/Controllers/Pages/TopController.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Pages;
 
-use App\Repositories\ArticleRepository;
+use App\Repositories\Article\FrontArticleRepository;
 use Illuminate\Contracts\View\View;
 use Illuminate\Routing\Controller;
 
 class TopController extends Controller
 {
     public function __construct(
-        private readonly ArticleRepository $articleRepository,
+        private readonly FrontArticleRepository $articleRepository,
     ) {}
 
     public function top(): View
