+++ tests/Feature/Repositories/ArticleRepository/GetLatestTest.php
@@ -8,13 +8,13 @@ use App\Enums\CategoryType;
 use App\Models\Article;
 use App\Models\Category;
 use App\Models\User;
-use App\Repositories\ArticleRepository;
+use App\Repositories\Article\FrontArticleRepository;
 use Illuminate\Database\Eloquent\Collection;
 use Tests\Feature\TestCase;
 
 class GetLatestTest extends TestCase
 {
-    private ArticleRepository $articleRepository;
+    private FrontArticleRepository $articleRepository;
 
     private User $user;
 
@@ -23,7 +23,7 @@ class GetLatestTest extends TestCase
     {
         parent::setUp();
         $this->user = User::factory()->create();
-        $this->articleRepository = app(ArticleRepository::class);
+        $this->articleRepository = app(FrontArticleRepository::class);
     }
 
     public function test(): void
