+++ tests/Feature/Repositories/ArticleRepository/PaginatePagesTest.php
@@ -7,13 +7,13 @@ namespace Tests\Feature\Repositories\ArticleRepository;
 use App\Enums\CategoryType;
 use App\Models\Article;
 use App\Models\User;
-use App\Repositories\ArticleRepository;
+use App\Repositories\Article\FrontArticleRepository;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Tests\Feature\TestCase;
 
 class PaginatePagesTest extends TestCase
 {
-    private ArticleRepository $articleRepository;
+    private FrontArticleRepository $articleRepository;
 
     private User $user;
 
@@ -22,7 +22,7 @@ class PaginatePagesTest extends TestCase
     {
         parent::setUp();
         $this->user = User::factory()->create();
-        $this->articleRepository = app(ArticleRepository::class);
+        $this->articleRepository = app(FrontArticleRepository::class);
     }
 
     public function test(): void
